[PopUp]モードセレクトのポップアップを修正。
